refactor(controller): extract component logic to service and model layers
Refactors the component controller into a Model-Service-Controller architecture to improve scalability and maintainability.
- Moves SQL queries to [src/models/component.model.js]
- Moves business logic and S3 integration to [src/services/component.service.js]
- Simplifies `src/controllers/component.controller.js` to handle only HTTP requests/responses
